# domains/core-platform-services/auth-service/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/contech/projects": {
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully"
          }
        },
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "ProjectsController_findAllProjects",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects retrieved successfully"
          }
        },
        "summary": "Get all projects",
        "tags": [
          "Projects"
        ]
      }
    },
    "/contech/projects/stats": {
      "get": {
        "operationId": "ProjectsController_getProjectStats",
        "parameters": [
          {
            "name": "managerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project stats retrieved successfully"
          }
        },
        "summary": "Get project statistics",
        "tags": [
          "Projects"
        ]
      }
    },
    "/contech/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findProjectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully"
          }
        },
        "summary": "Get project by ID",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully"
          }
        },
        "summary": "Update project details",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_removeProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          }
        },
        "summary": "Delete a project by ID",
        "tags": [
          "Projects"
        ]
      }
    },
    "/contech/projects/{id}/status": {
      "patch": {
        "operationId": "ProjectsController_updateProjectStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "status": "ACTIVE"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project status updated successfully"
          }
        },
        "summary": "Update project status",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/contech/tasks/project/{projectId}": {
      "get": {
        "operationId": "TasksController_findTasksByProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all tasks for a project",
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/contech/contracts/project/{projectId}": {
      "get": {
        "operationId": "ContractController_getContractsByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all contracts for a project",
        "tags": [
          "Contracts"
        ]
      }
    },
...
    "/contech/inspections/project/{projectId}": {
      "get": {
        "operationId": "InspectionController_findAllByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all inspections for a project with pagination",
        "tags": [
          "Inspections"
        ]
      }
    },
...
    "/contech/client-reports/projects/{projectId}": {
      "get": {
        "operationId": "ClientReportController_findAllByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all client reports for a project",
        "tags": [
          "Client Reports"
        ]
      }
    },
...
    "/contech/milestones/project/{projectId}": {
      "get": {
        "operationId": "MilestonesController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all milestones for a project",
        "tags": [
          "milestones"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Gateway documentation for frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the project",
            "example": "New Office Building"
          },
          "description": {
            "type": "string",
            "description": "Description of the project",
            "example": "Construction of 10-story office building"
          },
          "clientId": {
            "type": "string",
            "description": "Client ID associated with the project",
            "example": "client_12345"
          },
          "startDate": {
            "type": "string",
            "description": "Start date of the project",
            "example": "2025-01-01"
          },
          "endDate": {
            "type": "string",
            "description": "End date of the project",
            "example": "2025-12-31"
          },
          "budget": {
            "type": "number",
            "description": "Budget of the project",
            "example": 1000000
          },
          "location": {
            "type": "string",
            "description": "Project location",
            "example": "Addis Ababa"
          }
        },
        "required": [
          "name",
          "startDate"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "ACTIVE",
              "ON_HOLD",
              "DELAYED",
              "COMPLETED",
              "CANCELLED"
            ],
            "description": "Status of the project",
            "example": "ACTIVE"
          }
        }
      },
...
    }
  }
}